re-factor: lift shoppingList state up into App component
