+++ pages/edit.js
@@ -1,22 +1,9 @@
-import { useState } from "react";
-
 import Head from "next/head";
 import Header from "../components/Header";
 import NavigationBar from "../components/NavigationBar/NavigationBar";
 import ShoppingListEditor from "../components/ShoppingListEditor/ShoppingListEditor";
 
-import SHOPPING_LIST_DB from "../services/db.js";
-
-export default function Edit() {
-  const [editedShoppingListItems, setEditedShoppingListItems] =
-    useState(SHOPPING_LIST_DB);
-
-  function deleteItem(id) {
-    setEditedShoppingListItems((previousItems) =>
-      previousItems.filter((item) => item.id !== id)
-    );
-  }
-
+export default function Edit({ items, onDelete }) {
   return (
     <>
       <Head>
@@ -27,10 +14,7 @@ export default function Edit() {
 
       <main>
         <Header>Liste bearbeiten</Header>
-        <ShoppingListEditor
-          items={editedShoppingListItems}
-          onDelete={deleteItem}
-        />
+        <ShoppingListEditor items={items} onDelete={onDelete} />
         <NavigationBar />
       </main>
     </>
