+++ pages/index.js
@@ -1,24 +1,9 @@
-import { useState } from "react";
-
 import Head from "next/head";
 import Header from "../components/Header";
 import NavigationBar from "../components/NavigationBar/NavigationBar";
 import ShoppingList from "../components/ShoppingList/ShoppingList";
 
-import SHOPPING_LIST_DB from "../services/db.js";
-
-export default function Home() {
-  const [shoppingListItems, setShoppingListItems] = useState(SHOPPING_LIST_DB);
-
-  function toggleItemChecked(id) {
-    const updatedItems = shoppingListItems.filter((item) => item.id !== id);
-    const toggledItem = shoppingListItems.find((item) => item.id === id);
-    toggledItem.checked = !toggledItem.checked;
-    updatedItems.unshift(toggledItem);
-
-    setShoppingListItems(updatedItems);
-  }
-
+export default function Home({ items, onToggleItemChecked }) {
   return (
     <>
       <Head>
@@ -29,10 +14,7 @@ export default function Home() {
 
       <main>
         <Header>MyShoppingManager</Header>
-        <ShoppingList
-          items={shoppingListItems}
-          onToggleItemChecked={toggleItemChecked}
-        />
+        <ShoppingList items={items} onToggleItemChecked={onToggleItemChecked} />
       </main>
       <footer>
         <NavigationBar />
